+++ lib/eyes_appium.rb
@@ -25,9 +25,16 @@ if defined? Appium::Driver
   Appium::Core::Base::Driver.class_eval do
     def driver_for_eyes(eyes)
       if defined? Appium
-        Appium.promote_appium_methods Applitools::Appium::Driver::AppiumLib
+        Appium.promote_appium_methods(Applitools::Appium::Driver::AppiumLib)
       end
       Applitools::Appium::Driver.new(eyes, driver: self, is_mobile_device: true)
+    end
+  end
+
+  Appium::Driver.class_eval do
+    def driver_for_eyes(eyes)
+      Appium.promote_appium_methods(Applitools::Appium::Driver::AppiumLib, self)
+      Applitools::Appium::Driver.new(eyes, driver: self.start_driver, is_mobile_device: true)
     end
   end
 end